add fs

+++ kern/driver/ide.c
@@ -21,7 +21,7 @@ static struct ide_device ide_devices[MAX_IDE];
 
 
 static int
-ide_wait_ready(unsigned short iobase, bool check_error) {
+ide_wait_ready(unsigned int iobase, bool check_error) {
     return 0;
 }
 
@@ -39,12 +39,12 @@ ide_init(void) {
 }
 
 bool
-ide_device_valid(unsigned short ideno) {
+ide_device_valid(unsigned int ideno) {
     return VALID_IDE(ideno);
 }
 
 size_t
-ide_device_size(unsigned short ideno) {
+ide_device_size(unsigned int ideno) {
     if (ide_device_valid(ideno)) {
         return ide_devices[ideno].size;
     }
@@ -52,14 +52,14 @@ ide_device_size(unsigned short ideno) {
 }
 
 int
-ide_read_secs(unsigned short ideno, uint32_t secno, void *dst, size_t nsecs) {
+ide_read_secs(unsigned int ideno, uint32_t secno, void *dst, size_t nsecs) {
   if(CHECK_CALL(ideno, read_secs))
     return ide_devices[ideno].read_secs(&ide_devices[ideno], secno, dst, nsecs);
   return 0;
 }
 
 int
-ide_write_secs(unsigned short ideno, uint32_t secno, const void *src, size_t nsecs) {
+ide_write_secs(unsigned int ideno, uint32_t secno, const void *src, size_t nsecs) {
   if(CHECK_CALL(ideno, write_secs))
     return ide_devices[ideno].write_secs(&ide_devices[ideno], secno, src, nsecs);
   return 0;
